+++ src/CartridgeFixedFeeVanguard3v2.sol
@@ -5,10 +5,10 @@ import "./ICartridgeFeeModel.sol";
 
 contract CartridgeFixedFeeVanguard3v2 is ICartridgeFeeModel {
 
-    // uint128 constant cartridgeOwnerFee = 1000; // for 6 decimals
-    // uint128 constant protocolFee = 500; // for 6 decimals
-    uint128 constant cartridgeOwnerFee = 100000000000000; // for 18 decimals
-    uint128 constant protocolFee = 50000000000000; // for 18 decimals
+    // uint256 constant cartridgeOwnerFee = 1000; // for 6 decimals
+    // uint256 constant protocolFee = 500; // for 6 decimals
+    uint256 constant cartridgeOwnerFee = 100000000000000; // for 18 decimals
+    uint256 constant protocolFee = 50000000000000; // for 18 decimals
 
     function getMintFees(
         uint256,
