+++ src/TapeFixedFeeVanguard3v2.sol
@@ -5,12 +5,12 @@ import "./ITapeFeeModel.sol";
 
 contract TapeFixedFeeVanguard3v2 is ITapeFeeModel {
 
-    // uint128 constant cartridgeOwnerFee = 1000; // for 6 decimals
-    // uint128 constant tapeCreatorFee = 1000; // for 6 decimals
-    // uint128 constant protocolFee = 500; // for 6 decimals
-    uint128 constant cartridgeOwnerFee = 100000000000000; // for 18 decimals
-    uint128 constant tapeCreatorFee = 100000000000000; // for 18 decimals
-    uint128 constant protocolFee = 50000000000000; // for 18 decimals
+    // uint256 constant cartridgeOwnerFee = 1000; // for 6 decimals
+    // uint256 constant tapeCreatorFee = 1000; // for 6 decimals
+    // uint256 constant protocolFee = 500; // for 6 decimals
+    uint256 constant cartridgeOwnerFee = 100000000000000; // for 18 decimals
+    uint256 constant tapeCreatorFee = 100000000000000; // for 18 decimals
+    uint256 constant protocolFee = 50000000000000; // for 18 decimals
 
     function getMintFees(
         uint256 amount,
